+++ src/pages/rooms-overview.unit.js
@@ -1,5 +1,6 @@
 import { default as RoomsPage } from "./rooms-overview.vue";
 import RoomsModule from "@/store/rooms";
+import flushPromises from "flush-promises";
 
 const mockStoreData = [
 	{
@@ -27,7 +28,7 @@ const mockStoreData = [
 		shortTitle: "Ma",
 		displayColor: "#EC407A",
 		url: "/api/xxxx/1234w",
-		xPosition: 1,
+		xPosition: 4,
 		yPosition: 4,
 	},
 	{
@@ -36,8 +37,8 @@ const mockStoreData = [
 		shortTitle: "Bi",
 		displayColor: "#EC407A",
 		url: "/api/xxxx/1234w",
-		xPosition: 2,
-		yPosition: 4,
+		xPosition: 5,
+		yPosition: 5,
 		groupElements: [
 			{
 				id: "5",
@@ -60,8 +61,15 @@ const mockStoreData = [
 ];
 
 const spyMocks = {
-	storeRoomAlignMock: jest.spyOn(RoomsModule, "align"),
-	storeModuleFetchMock: jest.spyOn(RoomsModule, "fetch"),
+	storeRoomAlignMock: jest
+		.spyOn(RoomsModule, "align")
+		.mockImplementation(() => {}),
+	storeModuleFetchMock: jest
+		.spyOn(RoomsModule, "fetch")
+		.mockImplementation(() => {}),
+	storeModuleFetchAllMock: jest
+		.spyOn(RoomsModule, "fetchAllElements")
+		.mockImplementation(() => {}),
 	getElementNameByRefMock: jest.spyOn(RoomsPage.methods, "getElementNameByRef"),
 	openDialogMock: jest.spyOn(RoomsPage.methods, "openDialog"),
 	getDataObjectMock: jest.spyOn(RoomsPage.methods, "getDataObject"),
@@ -99,44 +107,35 @@ describe("RoomPage", () => {
 
 	it("should fetch the room data", async () => {
 		const wrapper = getWrapper();
-		await wrapper.vm.$nextTick();
-		await wrapper.vm.$nextTick();
+		await flushPromises();
 		expect(spyMocks.storeModuleFetchMock).toHaveBeenCalled();
-		expect(wrapper.vm.$data.roomsData).toStrictEqual(mockStoreData);
+		expect(wrapper.vm.items).toStrictEqual(mockStoreData);
 	});
 
-	it("should display 3 avatars component", async () => {
+	it("should display 6 avatars component", async () => {
 		const wrapper = getWrapper();
-		await wrapper.vm.$nextTick();
-		await wrapper.vm.$nextTick();
-		await wrapper.vm.$nextTick();
+		await flushPromises();
 		const avatarComponents = wrapper.findAll(".room-avatar");
-		expect(avatarComponents).toHaveLength(3);
+		expect(avatarComponents).toHaveLength(6);
 	});
 
 	it("should display 2 avatars component in 'mobile' device", async () => {
 		const wrapper = getWrapper("mobile");
-		await wrapper.vm.$nextTick();
-		await wrapper.vm.$nextTick();
-		await wrapper.vm.$nextTick();
+		await flushPromises();
 		const avatarComponents = wrapper.findAll(".room-avatar");
 		expect(avatarComponents).toHaveLength(2);
 	});
 
 	it("should display 1 group-avatar component", async () => {
 		const wrapper = getWrapper();
-		await wrapper.vm.$nextTick();
-		await wrapper.vm.$nextTick();
-		await wrapper.vm.$nextTick();
+		await flushPromises();
 		const groupAvatarComponents = wrapper.findAll(".room-group-avatar");
 		expect(groupAvatarComponents).toHaveLength(1);
 	});
 
 	it("should call 'openDialog' event if groupAvatar component clicked", async () => {
 		const wrapper = getWrapper();
-		await wrapper.vm.$nextTick();
-		await wrapper.vm.$nextTick();
-		await wrapper.vm.$nextTick();
+		await flushPromises();
 		const cardComponent = wrapper.find(".card-component");
 		await cardComponent.trigger("click");
 		expect(spyMocks.openDialogMock).toHaveBeenCalled();
@@ -144,43 +143,40 @@ describe("RoomPage", () => {
 
 	it("custom-dialog component should be visible", async () => {
 		const wrapper = getWrapper();
-		await wrapper.vm.$nextTick();
-		await wrapper.vm.$nextTick();
-		await wrapper.vm.$nextTick();
+		await flushPromises();
 		const cardComponent = wrapper.find(".card-component");
 		await cardComponent.trigger("click");
+		await flushPromises();
 		const customDialog = wrapper.find(".room-dialog");
+		await flushPromises();
 		const headline = customDialog.find("h2");
-		expect(customDialog.vm.isOpen).toBeTrue();
+		expect(customDialog.vm.isOpen).toBe(true);
 		expect(headline.element.innerHTML).toContain("Fourth");
 	});
 
 	it("should call the necessary methods for positioning while the page loading", async () => {
-		const wrapper = getWrapper();
-		await wrapper.vm.$nextTick();
-		await wrapper.vm.$nextTick();
+		getWrapper();
+		await flushPromises();
 		expect(spyMocks.getDataObjectMock).toHaveBeenCalled();
 		expect(spyMocks.findDataByPosMock).toHaveBeenCalled();
-		await wrapper.vm.$nextTick();
+		await flushPromises();
 		expect(spyMocks.getDeviceDimsMock).toHaveBeenCalled();
 		expect(spyMocks.hasGroupMock).toHaveBeenCalled();
 	});
 
 	it("'$refs' should be placed correctly for the components", async () => {
 		const wrapper = getWrapper();
-		await wrapper.vm.$nextTick();
-		await wrapper.vm.$nextTick();
-		await wrapper.vm.$nextTick();
+		await flushPromises();
 		expect(wrapper.vm.$refs["1-1"][0].$options["_componentTag"]).toStrictEqual(
 			"vRoomAvatar"
 		);
 		expect(wrapper.vm.$refs["2-2"][0].$options["_componentTag"]).toStrictEqual(
 			"vRoomAvatar"
 		);
-		expect(wrapper.vm.$refs["4-1"][0].$options["_componentTag"]).toStrictEqual(
+		expect(wrapper.vm.$refs["4-4"][0].$options["_componentTag"]).toStrictEqual(
 			"vRoomAvatar"
 		);
-		expect(wrapper.vm.$refs["4-2"][0].$options["_componentTag"]).toStrictEqual(
+		expect(wrapper.vm.$refs["5-5"][0].$options["_componentTag"]).toStrictEqual(
 			"vRoomGroupAvatar"
 		);
 		expect(wrapper.vm.$refs["3-3"][0].$options["_componentTag"]).toStrictEqual(
@@ -188,35 +184,32 @@ describe("RoomPage", () => {
 		);
 	});
 
-	it("should set the column count '2' if the device is 'mobile'", async () => {
+	it("should set the column count '4' if the device is 'mobile'", async () => {
 		const wrapper = getWrapper("mobile");
-		await wrapper.vm.$nextTick();
-		await wrapper.vm.$nextTick();
-		expect(spyMocks.getDeviceDimsMock).toHaveBeenCalled();
-		expect(wrapper.vm.dimensions.colCount).toBe(2);
-	});
-
-	it("should set the column count '4' if the device is 'tablet'", async () => {
-		const wrapper = getWrapper("tablet");
-		await wrapper.vm.$nextTick();
-		await wrapper.vm.$nextTick();
+		await flushPromises();
 		expect(spyMocks.getDeviceDimsMock).toHaveBeenCalled();
 		expect(wrapper.vm.dimensions.colCount).toBe(4);
 	});
 
-	it("should set the column count '6' if the device is 'desktop'", async () => {
-		const wrapper = getWrapper();
-		await wrapper.vm.$nextTick();
-		await wrapper.vm.$nextTick();
+	it("should set the column count '6' if the device is 'tablet'", async () => {
+		const wrapper = getWrapper("tablet");
+		await flushPromises();
 		expect(spyMocks.getDeviceDimsMock).toHaveBeenCalled();
 		expect(wrapper.vm.dimensions.colCount).toBe(6);
 	});
 
-	it("should call 'setDropElement' method for grouping after avatar-to-emptyAvatar drag&drop", async () => {
+	it("should set the column count '8' if the device is 'desktop'", async () => {
+		const wrapper = getWrapper();
+		await flushPromises();
+		expect(spyMocks.getDeviceDimsMock).toHaveBeenCalled();
+		expect(wrapper.vm.dimensions.colCount).toBe(8);
+	});
+
+	it("should call 'setDropElement' method after avatar-to-emptyAvatar drag&drop", async () => {
 		const wrapper = getWrapper();
 		const expectedPayload = {
 			from: {
-				x: 1,
+				x: 4,
 				y: 4,
 			},
 			item: {
@@ -225,27 +218,25 @@ describe("RoomPage", () => {
 				shortTitle: "Ma",
 				displayColor: "#EC407A",
 				url: "/api/xxxx/1234w",
-				xPosition: 1,
+				xPosition: 4,
 				yPosition: 4,
 			},
 			to: {
 				x: 4,
-				y: 4,
+				y: 2,
 			},
 		};
-		await wrapper.vm.$nextTick();
-		await wrapper.vm.$nextTick();
-		await wrapper.vm.$nextTick();
-		expect(wrapper.vm.$refs["4-1"][0].$options["_componentTag"]).toStrictEqual(
+		await flushPromises();
+		expect(wrapper.vm.$refs["4-4"][0].$options["_componentTag"]).toStrictEqual(
 			"vRoomAvatar"
 		);
 		expect(wrapper.vm.$refs["2-4"][0].$options["_componentTag"]).toStrictEqual(
 			"vRoomEmptyAvatar"
 		);
-		const avatarComponent = wrapper.findComponent({ ref: "4-1" });
+		const avatarComponent = wrapper.findComponent({ ref: "4-4" });
 		avatarComponent.trigger("dragstart");
 
-		const emptyAvatarComponent = wrapper.findComponent({ ref: "4-4" });
+		const emptyAvatarComponent = wrapper.findComponent({ ref: "2-4" });
 		emptyAvatarComponent.trigger("drop");
 
 		expect(spyMocks.setDropElementMock).toHaveBeenCalled();
@@ -262,33 +253,31 @@ describe("RoomPage", () => {
 		const expectedPayload = {
 			from: {
 				x: 1,
-				y: 4,
+				y: 1,
 			},
 			item: {
-				id: "3",
-				title: "Third",
+				id: "1",
+				title: "First",
 				shortTitle: "Ma",
-				displayColor: "#EC407A",
+				displayColor: "purple",
 				url: "/api/xxxx/1234w",
 				xPosition: 1,
-				yPosition: 4,
+				yPosition: 1,
 			},
 			to: {
 				x: 2,
 				y: 2,
 			},
 		};
-		await wrapper.vm.$nextTick();
-		await wrapper.vm.$nextTick();
-		await wrapper.vm.$nextTick();
-		expect(wrapper.vm.$refs["4-1"][0].$options["_componentTag"]).toStrictEqual(
+		await flushPromises();
+		expect(wrapper.vm.$refs["1-1"][0].$options["_componentTag"]).toStrictEqual(
 			"vRoomAvatar"
 		);
 		expect(wrapper.vm.$refs["2-2"][0].$options["_componentTag"]).toStrictEqual(
 			"vRoomAvatar"
 		);
 
-		const fromAvatarComponent = wrapper.findComponent({ ref: "4-1" });
+		const fromAvatarComponent = wrapper.findComponent({ ref: "1-1" });
 		fromAvatarComponent.trigger("dragstart");
 
 		const toAvatarComponent = wrapper.findComponent({ ref: "2-2" });
@@ -308,36 +297,34 @@ describe("RoomPage", () => {
 		const expectedPayload = {
 			from: {
 				x: 1,
-				y: 4,
+				y: 1,
 			},
 			item: {
-				id: "3",
-				title: "Third",
+				id: "1",
+				title: "First",
 				shortTitle: "Ma",
-				displayColor: "#EC407A",
+				displayColor: "purple",
 				url: "/api/xxxx/1234w",
 				xPosition: 1,
-				yPosition: 4,
+				yPosition: 1,
 			},
 			to: {
-				x: 2,
-				y: 4,
+				x: 5,
+				y: 5,
 			},
 		};
-		await wrapper.vm.$nextTick();
-		await wrapper.vm.$nextTick();
-		await wrapper.vm.$nextTick();
-		expect(wrapper.vm.$refs["4-1"][0].$options["_componentTag"]).toStrictEqual(
+		await flushPromises();
+		expect(wrapper.vm.$refs["1-1"][0].$options["_componentTag"]).toStrictEqual(
 			"vRoomAvatar"
 		);
-		expect(wrapper.vm.$refs["4-2"][0].$options["_componentTag"]).toStrictEqual(
+		expect(wrapper.vm.$refs["5-5"][0].$options["_componentTag"]).toStrictEqual(
 			"vRoomGroupAvatar"
 		);
 
-		const fromAvatarComponent = wrapper.findComponent({ ref: "4-1" });
+		const fromAvatarComponent = wrapper.findComponent({ ref: "1-1" });
 		fromAvatarComponent.trigger("dragstart");
 
-		const toAvatarComponent = wrapper.findComponent({ ref: "4-2" });
+		const toAvatarComponent = wrapper.findComponent({ ref: "5-5" });
 		toAvatarComponent.trigger("drop");
 
 		expect(spyMocks.addGroupElementsMock).toHaveBeenCalled();
@@ -353,8 +340,8 @@ describe("RoomPage", () => {
 		const wrapper = getWrapper();
 		const expectedPayload = {
 			from: {
-				x: 2,
-				y: 4,
+				x: 5,
+				y: 5,
 				groupIndex: 0,
 			},
 			item: {
@@ -364,7 +351,7 @@ describe("RoomPage", () => {
 			},
 			to: {
 				x: 4,
-				y: 4,
+				y: 5,
 			},
 		};
 
@@ -377,8 +364,8 @@ describe("RoomPage", () => {
 					shortTitle: "Bi",
 					displayColor: "#EC407A",
 					url: "/api/xxxx/1234w",
-					xPosition: 2,
-					yPosition: 4,
+					xPosition: 5,
+					yPosition: 5,
 					groupElements: [
 						{
 							id: "5",
@@ -401,16 +388,16 @@ describe("RoomPage", () => {
 			},
 		});
 
-		await wrapper.vm.$nextTick();
+		await flushPromises();
 		wrapper.vm.$refs.roomModal.$emit(
 			"drag-from-group",
 			wrapper.vm.groupDialog.groupData.groupElements[0]
 		);
 
-		await wrapper.vm.$nextTick();
+		await flushPromises();
 		expect(spyMocks.dragFromGroupMock).toHaveBeenCalled();
 
-		const emptyAvatarComponent = wrapper.findComponent({ ref: "4-4" });
+		const emptyAvatarComponent = wrapper.findComponent({ ref: "5-4" });
 		emptyAvatarComponent.trigger("drop");
 
 		expect(spyMocks.setDropElementMock).toHaveBeenCalled();
@@ -419,5 +406,46 @@ describe("RoomPage", () => {
 		expect(spyMocks.storeRoomAlignMock.mock.calls[0][0]).toStrictEqual(
 			expectedPayload
 		);
+	});
+
+	it("should search elements on dashboard", async () => {
+		const wrapper = getWrapper();
+		await flushPromises();
+
+		expect(wrapper.vm.$refs["1-1"][0].$options["_componentTag"]).toStrictEqual(
+			"vRoomAvatar"
+		);
+
+		const searchInput = wrapper.vm.$refs["search"];
+		searchInput.$emit("input", "thi");
+		await flushPromises();
+
+		expect(wrapper.vm.$refs["2-2"][0].$options["_componentTag"]).toStrictEqual(
+			"vRoomEmptyAvatar"
+		);
+		const avatarComponents = wrapper.findAll(".room-avatar");
+		expect(avatarComponents).toHaveLength(1);
+	});
+
+	it("should reset search text while dragging", async () => {
+		const wrapper = getWrapper();
+		await flushPromises();
+
+		expect(wrapper.vm.$refs["1-1"][0].$options["_componentTag"]).toStrictEqual(
+			"vRoomAvatar"
+		);
+
+		const searchInput = wrapper.vm.$refs["search"];
+		searchInput.$emit("input", "thi");
+		await flushPromises();
+		const avatarComponents = wrapper.findAll(".room-avatar");
+		expect(avatarComponents).toHaveLength(1);
+
+		const avatarComponent = wrapper.findComponent({ ref: "4-4" });
+		avatarComponent.trigger("dragstart");
+		await flushPromises();
+		const avatarComponentsAfterDragging = wrapper.findAll(".room-avatar");
+		expect(avatarComponentsAfterDragging).toHaveLength(6);
+		expect(wrapper.vm.$data.searchText).toStrictEqual("");
 	});
 });